services/workers/requirements.txt
praw==7.7.1
pytrends==4.9.2
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
boto3==1.34.0
psycopg2-binary==2.9.9
numpy==1.26.4
pandas==2.2.0
python-dotenv==1.0.1
pytest==8.0.0
